+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [4.4.1](https://github.com/BLSQ/openhexa-app/compare/4.4.0...4.4.1) (2026-03-03)
+
+
+### Bug Fixes
+
+* Also take into account backend prefix for static assets ([#1658](https://github.com/BLSQ/openhexa-app/issues/1658)) ([11f3db1](https://github.com/BLSQ/openhexa-app/commit/11f3db14ef654aa9107f0c2ef713052fadcaf335))
+
+## [4.4.0](https://github.com/BLSQ/openhexa-app/compare/4.3.0...4.4.0) (2026-03-03)
+
+
+### Features
+
+* add SubscriptionLimitTooltip and disable Workspace button at limit (HEXA-1524) ([#1643](https://github.com/BLSQ/openhexa-app/issues/1643)) ([5e33f62](https://github.com/BLSQ/openhexa-app/commit/5e33f6278cc453412aeed60b5e85e681e6585133))
+
+
+### Bug Fixes
+
+* stop sending  false to sentry no sample found after file preview of  non tabular files ([#1648](https://github.com/BLSQ/openhexa-app/issues/1648)) ([5c289a3](https://github.com/BLSQ/openhexa-app/commit/5c289a3b971804be507953518afb183beb4b2874))
+* Update GH action to allow adding prefix ENV vars... ([#1651](https://github.com/BLSQ/openhexa-app/issues/1651)) ([ef8ce7b](https://github.com/BLSQ/openhexa-app/commit/ef8ce7b59dbb8e792d913dff67f29fb7100776df))
+
 ## [4.3.0](https://github.com/BLSQ/openhexa-app/compare/4.2.0...4.3.0) (2026-02-26)